django-blog/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
defusedxml==0.6.0
distlib==0.3.0
dj-rest-auth==2.1.3
Django==3.0.4
django-allauth==0.44.0
django-cors-headers==3.2.1
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pycparser==2.20
pygame==2.0.1
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.10
Werkzeug==1.0.1
